Add files via upload
This version leaves more of the original dataset intact.  Instead of removing all users who have watched multiple movies on a single day 
(this makes resolving state, next state order impossible given only date information and not datetime information), we remove just the affected data. This means we remove tuples from all movies watched by the user on the day they binge-watched, as well as the tuple from the prior viewing session (for which we are unsure about the next state).